+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/assetProfileSmoke/CreateAssetProfileImportTest.java
@@ -49,8 +49,10 @@ public class CreateAssetProfileImportTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getAssetProfileByName(name) != null) {
-            testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+        if (name != null) {
+            if (getAssetProfileByName(name) != null) {
+                testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+            }
         }
     }
