+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/assetProfileSmoke/CreateAssetProfileTest.java
@@ -46,8 +46,10 @@ public class CreateAssetProfileTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getAssetProfileByName(name) != null) {
-            testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+        if (name != null) {
+            if (getAssetProfileByName(name) != null) {
+                testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+            }
         }
     }
