+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/assetProfileSmoke/SortByNameTest.java
@@ -43,8 +43,10 @@ public class SortByNameTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getAssetProfileByName(name) != null) {
-            testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+        if (name != null) {
+            if (getAssetProfileByName(name) != null) {
+                testRestClient.deleteAssetProfile(getAssetProfileByName(name).getId());
+            }
         }
     }
