+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/deviceProfileSmoke/CreateDeviceProfileTest.java
@@ -48,8 +48,11 @@ public class CreateDeviceProfileTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getDeviceProfileByName(name) != null) {
-            testRestClient.deleteDeviseProfile(getDeviceProfileByName(name).getId());
+        if (name != null) {
+            if (getDeviceProfileByName(name) != null) {
+                testRestClient.deleteDeviseProfile(getDeviceProfileByName(name).getId());
+            }
+            name = null;
         }
     }
