+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/ruleChainsSmoke/CreateRuleChainImportTest.java
@@ -52,8 +52,11 @@ public class CreateRuleChainImportTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getRuleChainByName(ruleChainName) != null) {
-            testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+        if (ruleChainName != null) {
+            if (getRuleChainByName(ruleChainName) != null) {
+                testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+            }
+            ruleChainName = null;
         }
     }
