+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/ruleChainsSmoke/RuleChainEditMenuTest.java
@@ -48,8 +48,11 @@ public class RuleChainEditMenuTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getRuleChainByName(ruleChainName) != null) {
-            testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+        if (ruleChainName != null) {
+            if (getRuleChainByName(ruleChainName) != null) {
+                testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+            }
+            ruleChainName = null;
         }
     }
