+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/ui/tests/ruleChainsSmoke/SearchRuleChainTest.java
@@ -43,8 +43,11 @@ public class SearchRuleChainTest extends AbstractDriverBaseTest {
 
     @AfterMethod
     public void delete() {
-        if (getRuleChainByName(ruleChainName) != null) {
-            testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+        if (ruleChainName != null) {
+            if (getRuleChainByName(ruleChainName) != null) {
+                testRestClient.deleteRuleChain(getRuleChainByName(ruleChainName).getId());
+            }
+            ruleChainName = null;
         }
     }
